moved the graphics files around again, awaiting future separation
