+++ g_src/graphics/renderer_curses.cpp
