+++ g_src/graphics/renderer_offscreen.cpp
